+++ fabric/src/main/java/com/faboslav/featurify/fabric/FeaturifyFabric.java
@@ -29,12 +29,12 @@ public final class FeaturifyFabric implements ModInitializer
 		}
 
 		RegistryManagerProvider.setRegistryManager(registryAccess);
-		LoadConfigEvent.EVENT.invoke(new LoadConfigEvent());
 		UpdateWorldgenDataEvent.EVENT.invoke(new UpdateWorldgenDataEvent(RegistryManagerProvider.getRegistryManager()));
 	}
 
 	private void onServerStart(MinecraftServer minecraftServer) {
 		RegistryManagerProvider.setRegistryManager(minecraftServer.registryAccess());
+		LoadConfigEvent.EVENT.invoke(new LoadConfigEvent());
 		UpdateWorldgenDataEvent.EVENT.invoke(new UpdateWorldgenDataEvent(RegistryManagerProvider.getRegistryManager()));
 	}
 }